Added FirewallAPI
TODO: Optimize

+++ Controller/Environments/Base.cs
@@ -91,16 +91,16 @@ namespace FlagDaemon.Controller.Environments
         Boolean PortListening(int PortNumber);
 
         // Return true if firewall allows inbound TCP connections to port from any IP
-        Boolean FirewallAllowsTcp(int PortNumber);
+        Boolean FirewallAllowsTcp(ushort PortNumber);
 
         // Return true if firewall allows inbound TCP connections to port from specific IPs
-        Boolean FirewallAllowsTcp(int PortNumber, List<String> AllowedIP);
+        Boolean FirewallAllowsTcp(ushort PortNumber, List<String> AllowedIP);
 
         // Return true if firewall allows inbound UDP connections to port from any IP
-        Boolean FirewallAllowsUdp(int PortNumber);
+        Boolean FirewallAllowsUdp(ushort PortNumber);
 
         // Return true if firewall allows inbound Udp connections to port from specific IPs
-        Boolean FirewallAllowsUdp(int PortNumber, List<String> AllowedIP);
+        Boolean FirewallAllowsUdp(ushort PortNumber, List<String> AllowedIP);
 
 
         /*
